+++ lib/jxl/decode.cc
@@ -2798,6 +2798,17 @@ JxlDecoderStatus JxlDecoderGetBoxSizeRaw(const JxlDecoder* dec,
   return JXL_DEC_SUCCESS;
 }
 
+JxlDecoderStatus JxlDecoderGetBoxSizeContents(const JxlDecoder* dec,
+                                              uint64_t* size) {
+  if (!dec->box_event) {
+    return JXL_API_ERROR("can only get box info after JXL_DEC_BOX event");
+  }
+  if (size) {
+    *size = dec->box_contents_size;
+  }
+  return JXL_DEC_SUCCESS;
+}
+
 JxlDecoderStatus JxlDecoderSetProgressiveDetail(JxlDecoder* dec,
                                                 JxlProgressiveDetail detail) {
   if (detail != kDC && detail != kLastPasses && detail != kPasses) {